create adapter and RecyclerView.ViewHolder

+++ app/src/main/java/com/fegusta/recyclerview/model/Jogo.kt
@@ -3,6 +3,6 @@ package com.fegusta.recyclerview.model
 data class Jogo(
         var codigo: Int,
         var nomeJogo: String,
-        var notaJogo: Int,
+        var notaJogo: Float,
         var console: String
 )
